Fully implemented position identification based on map and position pointers; started troubleshooting perlin noise repetition

+++ systemstest.h
@@ -71,7 +71,7 @@ tile pattern
 
 typedef struct
 {
-	float N;
+	double N;
 } POS;
 
 typedef struct
